Revert command structure back to `telegraf <component> <command>`.
